Now finally, I have the views communicating together
It did not work with SelectionService in Eclipse, because the
MultiPageEditor is already equipped with a SelectionProvider and there
can be only one. This selectionProvider skips all selections which are
not text and it also skips  if the activeEditor == null (i.e. the
resultsView is selected). Too bad ... now I had to couple the views
together tightly.