[#5] Generating java files from Xcore models should be done automatically during the build
Reverting all changes

This issue won't fix easily, because there seems to be no way to make it
work properly.

The Xcore model file needs a modelDirectory="..." statement to determine
where to place the generated files. However, when we set this directive
in the GenModel part, then the build completly failed with unmapped
paths and API breaks.

It is fully okay to check in the src-gen files, because then everybody
can use the model - even if they do not have Xcore installed.

Closes #5

Task-Url: http://github.com/roberthilbrich/assist-public/issues/5